Add models for Treatment subtypes
Treatments can be categorized into 7 subtypes.  Add a type field to the
Treatment model to track the category and use STI to keep all the
treatment types in the same database table.  Update factories and model
specs to support changes.

All currently existing Treatments will be migrated to be of type
Medicine.