Remove random seeding for tests
Tests break as a result of random seeding.  This implies
cross-dependency between tests as a result of the database being left
dirty as the result of a test.

Remove the random seeding for now, but acceptance tests should be
cleaned up such that randomizing the order should not break the tests.