Limit NPI search to taxonomy group_1
I think group_1 covers "Individual or Group (of Individuals)" and
excluds "Non-individual" records, which prevents us from later searching
for non-searchable NPI records.  Not entirely sure if this is correct,
but it works.

Also add exception if Npi#find fails and allow city and state to be set
in the AssociationsController.