Add in some nasty hacks to fix API calls
To get the old models working with the new controllers, I need to
install two hacks until the User model changes are in place.  First,
allow users to always manage other users; after the change, this will be
restricted to only associate instances.  Also, do not validate the
existance of the install_id for a User; to get around the disable
validation trick that the old controller was using, in the future this
validation will move to the Member model.