Fix compatibility with field clients
That use "disease" instead of "condition".  Introduce a few hacks here
an there to allow us to respond with the right naming convention when
appropriate.  Adjust specs checking database values to be looking at the
new model names.

The changes in this commit should be reverted after the next forced
client upgrade.