Attempt to fix out of memory error in RemoteEvents
Change log_all to be asynchronous so we don't deserialize the events
hash during the request transaction.  Limit the amount of instance
variable storage, instance variables need to be serialized and stored in
the delayed job queue.  Add user_id to RemoteEvent, because auth_token
will change after every login, save the user_id when it is available
when the RemoteEvent is created.