Add support for rendering empty consult with CSS
Add case for an empty consult in case the user has a consult with no
messages.  Clean up CSS for consult message previews and add some
nesting for a new CSS class.  In the future, the SCSS should be
reorganized to take advantage of nesting and inheritence to avoid
duplication.

Consult and Message cards still require additional formatting and copy.

[Finishes #57701042]