Enable Mayo terms of service in fullscreen content
Clicking the Mayo copywrite in fullscreen content should launch the
terms of service if it is available.

This triggers the NativeBridge with the open_content command which the
client does not currently support.  As a work around until support is
added, change the window.location.href to the URL to render the terms of
service.

[Finishes #59399258]