Move Card rendering logic to CardSerializer
Get it out of the controller and the model.  Invoke serializer in
controller to rendering the card correctly.  Improve and simplify the
view hierarchy for rendering full and preview cards.  Clean up and
standardize delegation of responsiblity to Card resources.  Add specs
for CardSerializer to check all current rendering cases.