Fix association deletion related to owned pairs
Because pairs are not deleted if the owner of the associate is the same
as the user/origin, we cannot rely on the dependent destroy to cause a
cascading teardown.  Skip the link by selectively destroying related
associations.  Add a before_validation callback to nullify the pair_id
in the event the pair was destroyed earlier in the transaction.