Only call Member state callbacks on state change
We want to only call a method on entry to a state, but we cannot be
guaranteed that the record will be created via transition, so we have to
use model callbacks.  Since we also want to only call on entry, check
that status_changed? is true before firing the callback.