Set ScheduledMessage welcome call reminder
When a member books a ScheduledPhoneCall, schedule a reminder for them.
If the member changes the time of the ScheduledPhoneCall, reschedule the
reminder.  Adjust :hold_scheduled_messages to only fire on state change
so we don't auto-hold our messages.

Note: That was painful.  StateMachine has a hard time doing multiple
transitions on the same model instance without a reload, so be sure to
reload between state changes.  Fix some broken specs by creating
supporting models instead of building them so they are persisted and
that the cascading save callbacks don't cause circular calls.