Stub specs that user CarrierWave and S3
Don't allow external connections, stub the storage for S3.
Unfortunately, when the storage is stubbed, the url method for a
CarrierWave image uploader will only give the local path to the image.
Further, the uploader cannot be reloaded by an ActiveRecord#reload and
only seems to be reloaded when a completely new instance of the model is
created.