Convert TimeOffset model to store minutes
Instead of time objects in time columns. I was having issues with
extracting the information correctly from the database because ruby was
adding an assumed date when converting from Time to DateTime.

Adjust naming to make methods a little more consistent. A little light
cleaning of the model and specs as well.